Remove dupe error
This already outputs a warning to console. No need for a big red error message in the discord window as well. This can be confusing for users also. A dev might set some elements or vars and not use them yet. But it still throws those errors and people think something is wrong.